% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_co2.R
\name{ac_get_co2}
\alias{ac_get_co2}
\title{Get data from indoorco2map.org}
\usage{
ac_get_co2(x = "web")
}
\arguments{
\item{x}{Character specifying which dataset to download one of c("web", "download", "transit")}
}
\value{
A dataframe with up to date co2 measurements
}
\description{
Get data from indoorco2map.org
}
\examples{
\dontrun{
# Get building data with one of three options
ac_df <- ac_get_co2()
ac_df <- ac_get_co2("web")
ac_df <- ac_get_co2("download")

# Get transit data
ac_df <- ac_get_co2("transit")
}

}
