% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_co2_download.R
\name{ac_get_co2_download}
\alias{ac_get_co2_download}
\title{Get data from indoorco2map.org - "download version"}
\usage{
ac_get_co2_download()
}
\value{
A dataframe with co2 building measurements
}
\description{
Get data from indoorco2map.org - "download version"
}
\details{
This function will download the version of building
data that is available using the "download" button on
the website
}
\examples{

#' \dontrun{
ac_df <- ac_get_co2_download()
}
}
