% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_co2_transit.R
\name{ac_get_co2_transit}
\alias{ac_get_co2_transit}
\title{Get transit data from indoorco2map.org}
\usage{
ac_get_co2_transit()
}
\value{
A dataframe with up to date co2 transit measurements
}
\description{
Get transit data from indoorco2map.org
}
\examples{
\dontrun{
ac_df <- ac_get_co2_transit()
}
}
