% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_co2_web.R
\name{ac_get_co2_web}
\alias{ac_get_co2_web}
\title{Get data from indoorco2map.org}
\usage{
ac_get_co2_web()
}
\value{
A dataframe with up to date co2 measurements - "web version"
}
\description{
Get data from indoorco2map.org
}
\details{
This function will download the version of building
data that is displayed on the website map
}
\examples{
\dontrun{
ac_df <- ac_get_co2_web()
}

}
