% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_co2_transit.R
\name{json_to_df}
\alias{json_to_df}
\title{Turn a co2 json into a dataframe}
\usage{
json_to_df(x)
}
\arguments{
\item{x}{A json}
}
\value{
a dataframe
}
\description{
Turn a co2 json into a dataframe
}
