% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_co2_transit.R
\name{range01}
\alias{range01}
\title{Create a vector of values evenly spaced between 0 and 1}
\usage{
range01(x, ...)
}
\arguments{
\item{x}{vector of values}

\item{...}{other parameters passed to min() or max()}
}
\value{
vector of values evenly spaced between 0 and 1 the same length as x
}
\description{
Create a vector of values evenly spaced between 0 and 1
}
