% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_co2_transit.R
\name{split_to_numeric}
\alias{split_to_numeric}
\title{Split a string seperated by ";" into numbers}
\usage{
split_to_numeric(x)
}
\arguments{
\item{x}{a string of numbers seperated by ";"}
}
\value{
a numeric vector
}
\description{
Split a string seperated by ";" into numbers
}
