% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_co2.R
\name{ac_get_co2}
\alias{ac_get_co2}
\title{Get data from indoorco2map.org}
\usage{
ac_get_co2()
}
\value{
A dataframe with up to date co2 measurements
}
\description{
Get data from indoorco2map.org
}
\examples{
\dontrun{
ac_df <- ac_get_co2()
}

}
