% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_unnest_longer.R
\name{ac_unnest_longer}
\alias{ac_unnest_longer}
\title{Unnest a co2 df list-column into rows}
\usage{
ac_unnest_longer(x)
}
\arguments{
\item{x}{a dataframe from ac_get_co2()}
}
\value{
a dataframe in long format
}
\description{
Unnest a co2 df list-column into rows
}
\examples{
ac_df <- ac_get_co2("download") |> 
ac_unnest_longer()
}
