% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_get_co2.R
\name{co2_to_numeric}
\alias{co2_to_numeric}
\title{Convert the co2 character array to a numeric list}
\usage{
co2_to_numeric(x)
}
\arguments{
\item{x}{a list containing measurements from json}
}
\value{
list of co2 records
}
\description{
Convert the co2 character array to a numeric list
}
